pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = ""
authors = [
    {name = "mateuskolap",email = "mateuskolap@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "alembic (>=1.16.4,<2.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy-declarative-extensions (>=0.16.2,<0.17.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "phonenumbers (>=9.0.12,<10.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
pytest-cov = "^6.2.1"
ruff = "^0.12.9"
taskipy = "^1.14.1"

[tool.poetry]
packages = [
    { include = "app" }
]

[tool.ruff]
line-length = 88
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT", "FAST"]
ignore = ["PLR0913"]

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
migrate = 'alembic upgrade head'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]